기간별 판매액 그래프 병합
