+++ src/test/java/creative/market/repository/query/OrderProductQueryRepositoryTest.java
@@ -11,6 +11,7 @@ import creative.market.repository.category.KindGradeRepository;
 import creative.market.repository.dto.BuyerTotalPricePerPeriodDTO;
 import creative.market.repository.dto.CategoryParamDTO;
 import creative.market.repository.dto.SellerAndTotalPricePerCategoryDTO;
+import creative.market.repository.dto.SellerPricePerPeriodDTO;
 import creative.market.repository.order.OrderRepository;
 import creative.market.repository.user.SellerRepository;
 import creative.market.service.OrderService;
@@ -273,6 +274,483 @@ class OrderProductQueryRepositoryTest {
         assertThat(result2).extracting("totalPrice").containsExactly(169000L, 133000L);
 
     }
+
+    @Test
+    @DisplayName("기간별 카테고리 전체 판매자 판매액, 카테고리 = kindGradeId 기준")
+    void allSellerTotalPriceByPeriodAndCategory1() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추-봄-상품 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 474L));
+
+        //식량작물-쌀-일반계-상품(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 432L));
+
+        // 채소류-상추-적-상품 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 492L));
+
+        //then
+        // 채소류-배추-봄-상품 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 2000L, 1000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀-일반계-상품(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 161000L, 161000L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 채소류-상추-적-상품 (2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 전체 판매자 판매액, 카테고리 = kindId 기준")
+    void allSellerTotalPriceByPeriodAndCategory2() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추-봄 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1662L, null));
+
+        //식량작물-쌀-일반계(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1613L, null));
+
+        // 채소류-상추-적(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1680L, null));
+
+        //then
+        // 채소류-배추-봄 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 6000L, 3000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀-일반계(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 201000L, 201000L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 채소류-상추-적 (2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 전체 판매자 판매액, 카테고리 = itemId 기준")
+    void allSellerTotalPriceByPeriodAndCategory3() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, 211, null, null));
+
+        //식량작물-쌀(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, 111, null, null));
+
+        // 채소류-상추(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, 214, null, null));
+
+        //then
+        // 채소류-배추 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 20000L, 10000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 201000L, 201000L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 채소류-상추(2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 전체 판매자 판매액, 카테고리 = itemCategory 기준")
+    void allSellerTotalPriceByPeriodAndCategory4() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //과일류-사과-홍옥-상품 결제 내역 추가
+        Product product1 = getProduct("사과상품2222222", 120000, "사과입니다.", 702L, productOwner2); // 과일류-사과-홍옥-상품
+
+        OrderProductParamDTO orderProductParam1 = new OrderProductParamDTO(1, product1.getId());
+        List<OrderProductParamDTO> orderParamList1 = addOrderProductParamDTO(orderProductParam1);
+        orderService.order(productBuyer.getId(), orderParamList1, buyerAddress);
+
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(200, null, null, null));
+
+        //식량작물(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(100, null, null, null));
+
+        // 과일류(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(400, null, null, null));
+
+        // 모든 상품(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result4 = orderProductQueryRepository
+                .findAllSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, null));
+
+        //then
+        // 채소류(2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 20000L, 10000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 201000L, 201000L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 과일류(2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 120000L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 모든상품(2달전 ~ 이번달)
+        assertThat(result4.size()).isEqualTo(3);
+        assertThat(result4).extracting("totalPrice").containsExactly(0L, 221000L, 331000L);
+        assertThat(result4).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 특정 판매자 판매액, 카테고리 = kindGradeId 기준")
+    void sellerTotalPriceByPeriodAndCategory1() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추-봄-상품 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 474L), productOwner1.getId());
+
+        //식량작물-쌀-일반계-상품(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 432L), productOwner1.getId());
+
+        // 채소류-상추-적-상품 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, 492L), productOwner1.getId());
+
+        //then
+        // 채소류-배추-봄-상품 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 2000L, 1000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀-일반계-상품(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 채소류-상추-적-상품 (2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 특정 판매자 판매액, 카테고리 = kindId 기준")
+    void sellerTotalPriceByPeriodAndCategory2() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추-봄 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1662L, null), productOwner1.getId());
+
+        //식량작물-쌀-일반계(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1613L, null), productOwner1.getId());
+
+        // 채소류-상추-적(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, 1680L, null), productOwner1.getId());
+
+        //then
+        // 채소류-배추-봄 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 6000L, 3000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀-일반계(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 채소류-상추-적 (2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 특정 판매자 판매액, 카테고리 = itemId 기준")
+    void sellerTotalPriceByPeriodAndCategory3() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        // 채소류-배추-고랭지-상품 결제 내역 추가
+        Product product1 = getProduct("고랭지상품2222222", 100000, "고랭지 배추입니다.", 477L, productOwner1); //채소류-배추-고랭지-상품
+
+        OrderProductParamDTO orderProductParam1 = new OrderProductParamDTO(2, product1.getId());
+        List<OrderProductParamDTO> orderParamList1 = addOrderProductParamDTO(orderProductParam1);
+        orderService.order(productBuyer.getId(), orderParamList1, buyerAddress);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류-배추 (2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, 211, null, null), productOwner1.getId());
+
+        //식량작물-쌀(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, 111, null, null), productOwner1.getId());
+
+        //then
+        // 채소류-배추 (2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 6000L, 203000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물-쌀(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+    }
+
+    @Test
+    @DisplayName("기간별 카테고리 특정 판매자 판매액, 카테고리 = itemCategory 기준")
+    void sellerTotalPriceByPeriodAndCategory4() throws Exception {
+        Address buyerAddress = createAddress("1122", "봉사산로", 12345, "동호수");
+        Seller productOwner1 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner1);
+        Seller productOwner2 = createSeller("강대현2", "111", "11", "19990212", "sd12fwf@mae.com", "010-3544-4444", createAddress("1111", "봉사산로", 12345, "1동1호"), "상호명1");
+        em.persist(productOwner2);
+
+        Buyer productBuyer = createBuyer("성호창32", "311111", "332222222", "19990512", "sdfw67f@mae.com", "010-3774-5555", buyerAddress);
+        em.persist(productBuyer);
+
+        sellerTotalPriceGraphInit(productOwner1, productOwner2, productBuyer);
+
+        //과일류-사과-홍옥-상품 결제 내역 추가
+        Product product1 = getProduct("사과상품2222222", 120000, "사과입니다.", 702L, productOwner1); // 과일류-사과-홍옥-상품
+
+        OrderProductParamDTO orderProductParam1 = new OrderProductParamDTO(1, product1.getId());
+        List<OrderProductParamDTO> orderParamList1 = addOrderProductParamDTO(orderProductParam1);
+        orderService.order(productBuyer.getId(), orderParamList1, buyerAddress);
+
+        // 채소류-배추-고랭지-상품 결제 내역 추가
+        Product product2 = getProduct("고랭지상품2222222", 100000, "고랭지 배추입니다.", 477L, productOwner1); //채소류-배추-고랭지-상품
+
+        OrderProductParamDTO orderProductParam2 = new OrderProductParamDTO(2, product2.getId());
+        List<OrderProductParamDTO> orderParamList2 = addOrderProductParamDTO(orderProductParam2);
+        orderService.order(productBuyer.getId(), orderParamList2, buyerAddress);
+
+        //when
+        YearMonth startDate = YearMonth.now().minusMonths(2);
+        YearMonth endDate = YearMonth.now();
+
+        String twoMonthsAgo = LocalDateTime.now().minusMonths(2).format(DateTimeFormatter.ofPattern("yyyy-MM")); //2달전
+        String oneMonthsAgo = LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM")); //1달전
+        String now = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM")); //이번달
+
+        // 채소류(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result1 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(200, null, null, null), productOwner1.getId());
+
+        //식량작물(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result2 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(100, null, null, null), productOwner1.getId());
+
+        // 과일류(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result3 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(400, null, null, null), productOwner1.getId());
+
+        // 모든 상품(2달전 ~ 이번달)
+        List<SellerPricePerPeriodDTO> result4 = orderProductQueryRepository
+                .findSellerTotalPricePerPeriodAndCategory(startDate, endDate, new CategoryParamDTO(null, null, null, null), productOwner1.getId());
+
+        //then
+        // 채소류(2달전 ~ 이번달)
+        assertThat(result1.size()).isEqualTo(3);
+        assertThat(result1).extracting("totalPrice").containsExactly(0L, 6000L, 203000L);
+        assertThat(result1).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        //식량작물(2달전 ~ 이번달)
+        assertThat(result2.size()).isEqualTo(3);
+        assertThat(result2).extracting("totalPrice").containsExactly(0L, 0L, 0L);
+        assertThat(result2).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 과일류(2달전 ~ 이번달)
+        assertThat(result3.size()).isEqualTo(3);
+        assertThat(result3).extracting("totalPrice").containsExactly(0L, 0L, 120000L);
+        assertThat(result3).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+
+        // 모든상품(2달전 ~ 이번달)
+        assertThat(result4.size()).isEqualTo(3);
+        assertThat(result4).extracting("totalPrice").containsExactly(0L, 6000L, 323000L);
+        assertThat(result4).extracting("date").containsExactly(twoMonthsAgo.toString(), oneMonthsAgo.toString(), now.toString());
+    }
+
+    private void sellerTotalPriceGraphInit(Seller productOwner1, Seller productOwner2, Buyer productBuyer) {
+        Address orderAddress = createAddress("1111", "봉사산로", 12345, "동호수");
+
+        Product product1 = getProduct("상품111", 1000, "상품입니다111", 474L, productOwner1); // 채소류-배추-상품
+        Product product2 = getProduct("상품222", 2000, "상품입니다222", 475L, productOwner1); // 채소류-봄-배추-봄-중품
+        Product product3 = getProduct("상품111", 3000, "상품입니다333", 477L, productOwner2); // 채소류-배추-고랭지-상품
+        Product product4 = getProduct("상품222", 4000, "상품입니다444", 480L, productOwner2); // 채소류 배추-가을-상품
+
+        for (int i = 0; i < 2; i++) { // 주문(이번달 ~ 1달전까지)
+            LocalDateTime now = LocalDateTime.now();
+
+            OrderProductParamDTO orderProductParam1 = new OrderProductParamDTO(i + 1, product1.getId());
+            OrderProductParamDTO orderProductParam2 = new OrderProductParamDTO(i + 1, product2.getId());
+            List<OrderProductParamDTO> orderParamList1 = addOrderProductParamDTO(orderProductParam1, orderProductParam2);
+
+            Long orderId = orderService.order(productBuyer.getId(), orderParamList1, orderAddress);
+            Order findOrder = orderRepository.findById(orderId).orElseThrow(() -> new NoSuchElementException("주문 내역이 존재하지 않습니다."));
+            findOrder.changeCreatedDate(now.minusMonths(i).withMinute(1));
+
+            OrderProductParamDTO orderProductParam3 = new OrderProductParamDTO(i + 1, product3.getId());
+            OrderProductParamDTO orderProductParam4 = new OrderProductParamDTO(i + 1, product4.getId());
+            List<OrderProductParamDTO> orderParamList2 = addOrderProductParamDTO(orderProductParam3, orderProductParam4);
+
+            Long orderId2 = orderService.order(productBuyer.getId(), orderParamList2, orderAddress);
+            Order findOrder2 = orderRepository.findById(orderId2).orElseThrow(() -> new NoSuchElementException("주문 내역이 존재하지 않습니다."));
+            findOrder2.changeCreatedDate(now.minusMonths(i).withMinute(2));
+        }
+    }
+
 
     private List<OrderProductParamDTO> addOrderProductParamDTO(OrderProductParamDTO... orderProductParamDTOS) {
         List<OrderProductParamDTO> orderParamList = new ArrayList<>();
